refactor(LottoPurchasedInfo) : 코드 가독성 개선을 위한 수정 - countResult 필드의 이름을 winningCountByRank로 변경 후 value 타입을 Long으로 변경 - groupByLottoRank 메서드의 가독성 개선 - calculateTotalRewards 메서드의 이름을 calculateTotalProfit으로 변경 - calculateTotalProfit의 총 수익금 계산 로직 수정 - calculateProfitPercentage 메서드의 코드 수정

+++ src/main/java/lotto/controller/LottoCheckerController.java
@@ -20,7 +20,7 @@ public class LottoCheckerController {
         List<Integer> winningNumbers = winningLotto.getWinningNumber();
         Integer bonusNumber = winningLotto.getBonusNumber();
 
-        Map<LottoRank, Integer> countResult = lottoPurchasedInfo.getWinningResult(winningNumbers, bonusNumber);
+        Map<LottoRank, Long> countResult = lottoPurchasedInfo.getWinningResult(winningNumbers, bonusNumber);
         Double profitPercentage = lottoPurchasedInfo.calculateProfitPercentage();
 
         return new LottoResultDTO(countResult, profitPercentage);
