+++ src/main/java/lotto/dto/LottoResultDTO.java
@@ -6,15 +6,15 @@ import lotto.model.LottoRank;
 
 public class LottoResultDTO {
 
-    private final Map<LottoRank, Integer> result;
+    private final Map<LottoRank, Long> result;
     private final Double profitPercentage;
 
-    public LottoResultDTO(Map<LottoRank, Integer> result, Double profitPercentage) {
+    public LottoResultDTO(Map<LottoRank, Long> result, Double profitPercentage) {
         this.result = result;
         this.profitPercentage = profitPercentage;
     }
 
-    public Map<LottoRank, Integer> getResult() {
+    public Map<LottoRank, Long> getResult() {
         return Collections.unmodifiableMap(result);
     }
